+++ frontend/src/composables/snap.js
